+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+- Logging interface
+- Specific login log api
+
 ### Changed
 
 - changed `UserJoinEvent(event_id)` to `UserJoinEvent((group_id, event_id))`