+++ CHANGELOG.md
@@ -7,13 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.1.6]
+
 ### Added
 
 - Logging interface
 - Specific login log api
-
-## Added
-
 - Added `decline_user_request_event_invite` update call
 - Added `decline_owner_request_event_invite` update call